BACKEND/requirement.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart

# Speech-to-text
openai-whisper

# ML backend
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2

# Speaker diarization
pyannote.audio==2.1.1

# HuggingFace
transformers
huggingface_hub

# NLP
keybert
scikit-learn
sentencepiece

# Utilities
ffmpeg-python